Avanzo en la vista. Utilizo AJAX con Jquey para actualizar selects

+++ src/main/java/com/mihuella/repositories/OrganizacionRepo.java
@@ -3,5 +3,6 @@ package com.mihuella.repositories;
 import com.mihuella.organizacion.Organizacion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrganizacionRepo extends JpaRepository<Organizacion, Long> {
+public interface OrganizacionRepo extends JpaRepository<Organizacion, Integer> {
+
 }
